Move page objects to top level folder within acceptance tests
Previously page objects have been nested within feature folders. The
problem that can arise here is that they are all on the `client.page`
namespace and if pages are given the same filename they will override
any previous pages with that name.

By moving them to their own space it is also easier to see top level
pages that are shared by other pages or that can be extended to provide
a shared level of elements and commands across all pages.

The folder has been called `pages/` as this is inline with how
Nightwatch namespaces once they are loaded onto the client even
though in some documentation they are referred to as _page objects_.
