+++ test/acceptance/features/companies/step_definitions/details.js
@@ -1,7 +1,7 @@
 const { client } = require('nightwatch-cucumber')
 const { Then } = require('cucumber')
 
-const Company = client.page.Company()
+const Company = client.page.companies.Company()
 
 Then(/^the company details UK region is displayed$/, async function () {
   const { ukRegion } = this.state.company
