+++ test/acceptance/pages/events/List.js
@@ -2,7 +2,7 @@ const {
   getSelectorForElementWithText,
   getButtonWithText,
   getMetaListItemValueSelector,
-} = require('../../../helpers/selectors')
+} = require('../../helpers/selectors')
 
 const getFilterTagRemoveBtnSelector = (text) => getSelectorForElementWithText(
   text,
